Remove RRAttention v1 support
